Resolved the issues raised
