+++ tos_custom/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TosCustomConfig(AppConfig):
+    name = 'tos_custom'
